Disabled check for wired connection and automatic hotspot enable.
For my current use case there are instances where I need to have the wired connection and
a client WLAN connection instead of automatically enabling hotspot. Only if the configured
wireless networks are not available do I want the hotspot mode, regardless if there is a
wired connection.

Future investigation would be to make it a configuration value sourced from a configuration
file, e.g. in /etc/default.